Added Ondřej Surý GPG key ID for Debian repo
This commit adds the new key ID for the Debian repo of Ondřej Surý
to ensure the GPG key won't be downloaded by the apt_key module if
it's already installed on the system.

This modification is made for systems behind several firewalls or
filtering proxies on which the key is pre-deployed and when the
Ansible module doesn't work well with proxy env vars (eg. apt_key)